Align header, footer
The body was extended to encompass the full screen, and was turned into a flex element.
With the justify-content property and its value space-between, the header was aligned to the
top, and the footer to the bottom. Moreover, the footer content was centered with addtitional
flex properties. The head was turned into a flex element as well such that the logo and the links
are spaced apart again with the justify-content property like above.

Moreover, the links in the header were stripped of margin, padding, and usual link decorations.

Finally, I added the divs for the page content. According to the reference image, there are going to be
four sections: .page-title, .page-info, page-quote, page-sign-up.